Fever dream restructuring of Homelab work
Put old Nomad specific config in a folder. I started a new folder
to hold config to run a VM that would host pihole, cloudflared, and
traefik, all running as docker containers. All of this config was
put into two distinct Terraform configurations (need to do this because
the Docker provider can't wait for the VM to be created - annoying).